Add tirprog program gluing 2 pieces of code together.
This code is largly the work of others, simply glued together.

First, my TIRSetter arduino program, which has credits in the readme
here:
https://github.com/Materdaddy/TIRSetter

Second, the USBDMXPro code from FPP here:
https://github.com/FalconChristmas/fpp

This program is to re-program your TIR flood's start address.  This code
has been tested as working using a counterfeit FTDI-based USB-to-RS485
dongle obtained for a few bucks on eBay on both a Debian 7.0 system as
well as a raspberry pi.